Fix #113 (deactivate confirmation token, kind of)
It was already okay as a new token was generated, invalidating the old URL.
However, I don't see a reason why we actually need a new token in the database,
so we just empty the field. This way, we know when a field is empty that the
team was already activated successfully. If they are not active afterwards, it
must have been manually.